+++ app/controllers/supply_teachers/admin/suppliers_controller.rb
@@ -1,0 +1,7 @@
+module SupplyTeachers
+  module Admin
+    class SuppliersController < SupplyTeachers::Admin::FrameworkController
+      include ::Admin::SuppliersController
+    end
+  end
+end