+++ app/controllers/supply_teachers/rm6238/admin/lot_data_controller.rb
@@ -1,0 +1,8 @@
+module SupplyTeachers
+  module RM6238
+    module Admin
+      class LotDataController < SupplyTeachers::Admin::LotDataController
+      end
+    end
+  end
+end