+++ features/step_definitions/accessibility_steps.rb
@@ -1,0 +1,3 @@
+Then('the page should pass the accessibility checks') do
+  step 'the page should be axe clean excluding ".ccs-contact-us, .govuk-phase-banner"'
+end